Eliminates race conditions, still rigorous testing to be done.
This code requires a couple of optimizations in the dtrace probe
enabling and disabling code as it's 3 times as slow as it was
originally. While perhaps this is not the most critical thing to be
worrying about right now, a couple of simple optimizations can be made
to make this a tad more efficient.